refactor: GameController, Utils 각각 패키지 만들어 위치 이동
