+++ src/main/java/racingcar/Application.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import java.util.Scanner;
 
+import racingcar.controller.GameController;
+
 public class Application {
 	public static void main(String[] args) {
 		Scanner scanner = new Scanner(System.in);
