+++ src/main/java/racingcar/controller/GameController.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import racingcar.domain.Cars;
 import racingcar.validator.MoveCountValidator;
