+++ src/test/java/racingcar/validator/RacingGameValidatorTest.java
@@ -6,7 +6,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import racingcar.Utils;
+import racingcar.Util.Utils;
 
 class RacingGameValidatorTest {
 	@Test
